perf: 이미지 크롭 & 프롬프트 코드
